requirements.txt
gdown
requests
tensorflow
streamlit
pillow
numpy